feat(web): add independent routes for settings sub-pages and redesign sidebar
- Add individual routes for settings sub-pages (/settings/profile, /settings/account, etc.)
- Refactor settings page into modular components under pages/settings/
- Redesign sidebar navigation with OpenClaw-style collapsible sections
- Update URL when switching tabs for better navigation and refresh support
- Extract components: ProfileSettings, AccountSettings, AppearanceSettings,
  NotificationSettings, InviteSettings, TaskSettings, QuickstartSettings
